+++ mapjs/src/core/content/auto-themed-idea-utils.js
@@ -25,7 +25,7 @@ const calcIdeaLevel = require('./calc-idea-level'),
 	},
 	recalcIdeasAutoNodeAttrs = (activeContent, themeObj, idea, level, numberOfSiblings) => {
 		'use strict';
-		// console.debug('recalcIdeasAutoNodeAttrs idea.id', idea.id, 'level', level, 'numberOfSiblings', numberOfSiblings); //eslint-disable-line
+		Logger.debug('recalcIdeasAutoNodeAttrs idea.id', idea.id, 'level', level, 'numberOfSiblings', numberOfSiblings); //eslint-disable-line
 		if (level > 0) {
 			recalcAutoNodeAttrs(activeContent, themeObj, idea, level, numberOfSiblings);
 		}